(maint) Define vardir when running tests
This is an odd one.

When running tests with the puppet 4 gem (but not puppet 5 or puppet 6),
this test fails with the following error:
```
     Errno::ENOTDIR:
       Not a directory @ dir_s_mkdir - /dev/null/state/transactionstore.yaml20190109-6277-sfid86.lock
```

From jumping into pry and poking around, `Puppet[:vardir]` is set to
`/dev/null`
```
[1] pry(#<RSpec::ExampleGroups::PuppetTransaction>)> Puppet[:vardir]
=> "/dev/null"
[2] pry(#<RSpec::ExampleGroups::PuppetTransaction>)> Puppet[:statedir]
=> "/dev/null/state"
[3] pry(#<RSpec::ExampleGroups::PuppetTransaction>)> Puppet[:transactionstorefile]
=> "/dev/null/state/transactionstore.yaml"
```

Consequently, `/dev/null/state` doesn't exist, so when we try to manage
`transactionstore.yaml` in that path, we get the above failure. Rather
than create a directory in `/dev/null`, it makes more sense to create it
in a temporary directory that we can clean up later.

The frustrating this is that this particular quirk seems to be isolated
to puppet 4. There's nothing obvious that can point to why, so in the
meantime, it seems this hacky fix is our best bet.